Add versioning, attestation, and release assets (#2)
This adds versioning using changesets (private packages) so that we can have each firmware build handled differently.

Changesets, adds tags and releases to the repository so we can use that to trigger workflows hence perform provenance attestation and upload as release artifacts.

Zephyr has inbuilt application versioning logic we only need to generate the VERSION file which is done using a javascript script.