Add WiFi connection for internet access (#6)
Add support for WiFi connection on both boards. However, the ESP32S3 seems to not support EAP because there are build errors. Also, we have credentials "hard-coded" which should not be the case in production but this moves the needle.

To get EAP to work, the Zephyr RTOS version referenced in not `v4.1.0` but a recent commit where changes by NXP exist. This is reflected in `west-manifest/west.yml`

- Created TROUBLESHOOTING.md for common issues and solutions.
- Enhanced Kconfig and prj.conf for WiFi settings, including enterprise authentication.