Major improvements to internet connectivity
I must note that the WiFi on zephyr is really bad. So many quirks. When you setup connection manager, you loose `NET_EVENT_WIFI_DISCONNECT_RESULT` events. On RW612, reconnection happens automatically. On ESP32, you need to reconnect. Relying on connection manager to do retries does not work on RW612 because it has its own retries. On ESP32, it also fails with error `-EALREADY` and if you issue a disconnect command it does not work.

It is likely that how I am setting things up is where the problem might be but it has taken me a week to figure out connectivity which is not very encouraging, let alone the circus that is enterprise WiFi. EAP requires `CONFIG_WIFI_NM_WPA_SUPPLICANT=y` which results in a broken build for ESP32 and for RW612, you need a lot more configuration to get it to work which I did.

WiFI relocated to SRAM on the ESP32 results in a panic crash.