Added platformio-based firmware
This is meant to catapult the project since so much already works with Arduino/ESP32. With Zephyr, there is still a lot of things that do not work with WiFi, so much that I cannot connect to basic WiFi networks and get tracffic. For enterprise, it is another mess of s story because HostAp/WpaSupplicant is still quite experimental.

The zephyr ones remains because it is meant to the be the eventual thing we use because zephyr has much better primitives (logging, events, work queues, credentials, etc) and drivers (no need to write own). With time, its WiFi will mature enough. It might be that I am not as experienced with it because I suspect some people are using this WiFi stack in production.

Another alternative is using ESP-IDF but since we have a simple arduino thing working, there is no need to do so.

This base code already works with WiFi and mDNS, shortly cloud support will be added and the sensors logic completed.