Added device twin logic to house firmware version and update info
Firmware updates are a requirement for the project. The easiest way to do this is via device twin properties in Azure IoT Hub where we set the current one in reported properties and receive updates via desired props. This PR adds logic for working with the device twin in this aspect. It is very specific to the use case of this project both in simplicity and structure.
To make this logic more appropriate the versioning logic in `version.py` has been updated to match that for zephyr. This means we have both a string version (semver) and an integer version (shall be used to check if newer).

Actual firmware updates logic will follow.