Add actuator functionality which syncs with device twin (#16)
Implement actuator class with configuration and state management that works with the cloud hub for device twin synchronisation. The actuator config comes from the device twin's desired properties while the actuator state is pushed to the device twin's reported properties.

On reboot all the state is lost so the reported properties are reset back to zero. While this can be changed to store in `Preferences`/`NVM`, there is no real need to do so at this time.