Only recommend nRF Connect extensions hence nest west.yml
The Zephyr IDE extension my `mylonics` does not support macOS correctly. I got it to work once but can't reproduce on a different computer. The  other extension (AC6 Workbench) doesn't work for me either so the solution is to use the nRF Connect extension(s). The `west.yml` file can not be safely placed in the `firmware` folder because it must not be in a specific location as required by previous extension.

Unfortunately, NCS does not support esp32s3  out of the box (requires xtensa toolchain) and still haven't figured out how to add that support.

By and large the best way to go about this is to stick to the official west cli but