Add support for ESP32-C6 DevKitC (#63)
The ESP32C6 has support for TEE which is analogous to T-FM on the ARM side of things. It is a better board to validate the security concerns that are core in this project.
Consequently, when working with nRF Connect SDK (NCS), I have not figured out how to enable Xtensa in the toolchain or to download another toolchain where it exists. However, Nordic is now using RISC-V cores with nRF54L and nRF54H series and this is also what ESP32C6 is based on. I can now build for ESP32 based boards from within NCS.

Overall, this is a learning for me to work with RISC-V.